requirements.txt
streamlit
faiss-cpu
numpy
sentence-transformers
openai
scikit-learn
